# docs/nus-nextbus-openapi.yaml
openapi: "3.0.0"
info:
  title: NUS NextBus
  description: "Unofficially reverse-engineered API used by NUS NextBus application."
  version: 2.0.0
servers:
  - url: https://nnextbus.nus.edu.sg
    description: Current server used by NUS NextBus app
paths:
...
  /ShuttleService:
    get:
      summary: Get shuttle service at a stop
      operationId: GetShuttleService
      parameters:
        - in: query
          name: busstopname
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ShuttleServiceResult:
                    type: object
                    properties:
                      name: { type: string }
                      caption: { type: string }
                      shuttles:
                        type: array
                        items:
                          type: object
                          properties:
                            name: { type: string }
                            arrivalTime: { type: string }
                            arrivalTime_veh_plate: { type: string }
                            nextArrivalTime: { type: string }
                            nextArrivalTime_veh_plate: { type: string }
                            passengers: { type: string }
                            nextPassengers: { type: string }
...
components:
  securitySchemes:
    Basic:
      type: http
      scheme: basic
security:
  - Basic: []